Website/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.9.0
django-versatileimagefield==2.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
numpy==1.23.3
opencv-python==4.5.3.56
Pillow==8.3.2
psycopg2-binary==2.9.1
python-magic==0.4.24
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
opencv-python-headless==4.5.3.56